requirements.txt
# Herramientas de Automatización - Dependencias

# Interfaz y Menú
rich
questionary

# Web Scraping (Monitor de Precios)
requests
beautifulsoup4

# Procesamiento de Documentos e IA (Resumidor)
google-genai
pypdf
python-dotenv

# Manipulación de Imágenes (Renombrador)
Pillow

# Utilidades Varias
yt-dlp
